[SERVINF-304] Added internal tracking statistics
Added statistics to show the some statistics about users of the biovel portal. This includes:
 a) the number of users that have registered, and
b) the number of active users (i.e. users who have ran a workflow)

These statistics are generated for each month since the creation date of the first user to the month that today is in.