[SERVINF-352] Fixed bug about the displaying of error message in the person details form.
In a previous commit, the render call had been changed to a redirect call because of a routing error. The redirect meant the previous form details weren’t being passed back hence no errors.
We found the problem with the original render in the headers. An if statement was assuming that a successfully created person would have current_user.person == true but this is true for the failed person too.
Another check to make sure the current_user.person isn’t a new_record has been added.