Add rake tasks to start and stop delayed job workers.
If workflows are enabled it also starts Taverna Player workers, the number of
which is specified as a parameter to the task, e.g.:

  rake seek:workers:start[2]

will start 2 Taverna Player workers and a SEEK worker. If the parameter is
omitted or set to zero no Taverna Player workers will be started. A SEEK worker
is always started.

Other tasks provided are:

  rake seek:workers:status
  rake seek:workers:stop

which operate on all running workers.