adding a model with multiple files or url updated
fixes to the parameters the get passed, and also a reworking of the javascript for handling the added files and urls. Fixes some problems
in the earlier version that were discovered to do with filename and url not being in sync, and the flag to make a copy not being carried across.
Handling listing the files for a new version still needs updating, which will be doen once things are tidied up.