Refactored out unessary code in filtering
included special cases for filtering assays and studies for project_and_descendants. This was unnessary since this method can
be delegated though to the invesitigation via the already delegated projects method.
Also moved the include ItemsProjectsExtension over the acts_as_isa to ensure it is included in all ISA models.