fixed the configurations and test configuration, removed forced_default
put back the configurations to what they should be. the forced_default was confusing and implied it behaved the same as fixed. The configurations supported by the admin pages is not a concern of the general Seek::Config class and can be documented elsewhere or by groupings in the seek_configuration file.
Fixed profile_select_by_default with seems to have been changed by a search and replace.