ajax callback to get and display data file preview - OPSK-20
ajax fetches json details about datafile and displays it in a preliminary preview. Ajax is triggered when the select box is changed.
Preview is cleared when no data file is selected. The preview can easily be changed as it is handled javascript function and displayed according to a shared partial.