project admin button changed to take straight to admin members
the old admin page didn't haven't anything else of much use, and just required another click to get to the relevant page.
changed the button path and text to reflect this.
the old admin page remains in place, should we need it to reintroduce the updating of JERM settings in the future.