fixed tests that weren't web-mocked out properly
overzealous exception handling in content-blob was convering this up. Since being fixed a couple of tests were failind due to http calls not being mocked out with webmock (since the exception was previously being caught).
Now changed to mock out the http calls with a 404 error, which will produce the same error expected in the tests