fix to failing project subscription test
a subscription with the same hardcoded id already existed, which looks to be due to the addition of a faceted browsing test that creates a bunch of assets.
now changed to use a project subcription id that is less likely to preexist, and also attempt to destroy it if it does.