should be fix to failing project subscription test!
it was relying that the order the person.project_subcriptions were returned always matched the order to match the subcription project id, which isn't necessarily the case. Now updated to avoid relying on the order.
Also fixed how the params are passed, and added an additional check that the subcriptions are deleted