OPSK-197 - fix to javascript error creating model
caused when selecting a model format or type. The javascript related to some redundant
functionality that allowed you to remove a model type or format. This functionality no longer exists so the javascript has now been updated to fix this.
The functionality to self manage these types is likely to be re-added as part of FAIRDOMHub - but it's behaviour will be different to reflect a larger and more diverse userbase and programme structures.