OPSK-211 hide contact details outside of programme
only show contact details like email, skype for people that share the same programme.
as a side effect, the distiguishing info for author lists has been changed from email to their projects, as the decision to display each email is has become more complex and needs testing for each person in relation to the current person.