experimenting with excluding from codeclimate
a couple of files that inadvertantly affect the score but we are currently unable to fix at the moment.
modified_cytoscape.js - fixes and extends some cytoscape stuff, and we don't want to diverge from the original too much. hoping to update to more recent version soon and remove this file
bio_samples.rb - loads of copy&pasted code and massive methods - only used by VLN and likely to be deprecated later this year. Not worth fixing.