test fixes
had to put back User.current_user in a policy_controller method. It wasn't a problem itself, but due to the way some PolicyControllerTests currently work - calling the controller directly rather than invokign a request. This prevents the @request and session being correctly setup. Currently not the time to update the test, so reverted and added a some FIXME notes

also updated site_announcements gem to reflect removal of currently_logged_in? method which wasn't used elsewhere in SEEK.