OPSK-959 added failing test
to cover the case when something is made public, but has a permission to restrict for a certain user. The user can then see the thing when logged out, but not when logged in. Added additional case where there is a less restrictive permission included for a another user, as we don't want the fix to skip these less restrictive permissions entirely.