OPSK-1287 - correctly honour the ALL_USERS scope
correct a mistaken assumption we could treat this as any user, including anonymous. It's an old sharing policy that is now redundant since anybody can log it, but for older versions of SEEK is should still be honoured.
In a future update, we are considering removing this field entirely and converting these items to private