OPSK-696 updated the has_many relationships between Person and various types
updates to the link for <type>_for_person, to no longer be polymorphic.
There was a previous reliance on a link between User and items for some
related items. Will move on to changing these associations to just
person.type (e.g person.studies) and include all types, in a future
change