OPSK-2030 some initial work for a generic "Add" button for items
a button that allows you to add a new item, automatically assigning the
current item in the form.
attempting to generate a generic helper for the button, with a
definition defining what can be added and what the parameters should be.
the link is passed using standard parameters that match the form, and
what would be provided when creating/updating the item