Bug fix in the publ. creation, improvements and cleanups.
uploadPdf was too late in the create method and was preventing other data from being saved. Various improvement and cleanup. Big cleanup of migrations, but maybe the version ID is also not needed in publications.
Added misc_link to AssetLink in preparation for misc URLs in publications.

Uploaded file are still visible in the list of publications when the uploading is not allowed, on purpose, so we know that some files are there.