Fixed the cache approach, fixed the policy. Added tests.
The cache was also caching the blobs, so could break depending on the blob. Replaced by caching the class name so the detect does not have to be executed at each call (idea and code from Finn, thanks!)
Uses allow_embedding instead of options == ['necessary'] as the consent might not be needed.
Added integration tests for the embedded content.
Small bug fixes.