fixed an issue where the original implementation of DESCRIBE was ignored due to blank.
What happens now is if you run a sparql query and click on any of the URLS's it will execute a DESCRIBE such as  `DESCRIBE <http://localhost:3000/projects/1>`

Why? This allows users to more easily browse through the graph. TODO, a graphical map to be added once the "test data" has been added and is covering all RDF aspects.